moving code about and using listview not datacell
